requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dataclasses-json==0.5.9
Flask==2.3.3
frozenlist==1.4.0
fsspec==2023.9.0
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
langchain==0.0.279
langsmith==0.0.33
llama-index==0.8.16
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.7
nltk==3.8.1
numexpr==2.8.5
numpy==1.25.2
openai==0.28.0
packaging==23.1
pandas==2.1.0
pydantic==2.3.0
pydantic_core==2.6.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.20
tenacity==8.2.3
tiktoken==0.4.0
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.7
yarl==1.9.2
